+++ CSK_Module_MultiDeepLearning/scripts/ImageProcessing/MultiDeepLearning/MultiDeepLearning_Controller.lua
@@ -214,7 +214,7 @@ Script.serveFunction("CSK_MultiDeepLearning.pageCalled", pageCalled)
 
 local function setInstance(dnnInstance)
   selectedInstance = dnnInstance
-  _G.logger:info(nameOfModule .. ": New selected instance = " .. tostring(selectedInstance))
+  _G.logger:fine(nameOfModule .. ": New selected instance = " .. tostring(selectedInstance))
   multiDeepLearning_Instances[selectedInstance].activeInUI = true
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'activeInUI', true)
   pageCalled()
@@ -227,7 +227,7 @@ end
 Script.serveFunction("CSK_MultiDeepLearning.getInstancesAmount", getInstancesAmount)
 
 local function addInstance()
-  _G.logger:info(nameOfModule .. ": Add instance")
+  _G.logger:fine(nameOfModule .. ": Add instance")
   table.insert(multiDeepLearning_Instances, multiDeepLearning_Model.create(#multiDeepLearning_Instances+1))
   Script.deregister("CSK_MultiDeepLearning.OnNewValueToForward" .. tostring(#multiDeepLearning_Instances) , handleOnNewValueToForward)
   Script.register("CSK_MultiDeepLearning.OnNewValueToForward" .. tostring(#multiDeepLearning_Instances) , handleOnNewValueToForward)
@@ -236,7 +236,7 @@ end
 Script.serveFunction('CSK_MultiDeepLearning.addInstance', addInstance)
 
 local function resetInstances()
-  _G.logger:info(nameOfModule .. ": Reset instances.")
+  _G.logger:fine(nameOfModule .. ": Reset instances.")
   setInstance(1)
   local totalAmount = #multiDeepLearning_Instances
   while totalAmount > 1 do
@@ -252,26 +252,26 @@ local function setModelByName(modelName)
   local model = Object.load(multiDeepLearning_Instances[selectedInstance].parameters.modelPath .. modelName)
 
   if model then
-    _G.logger:info(nameOfModule .. ": Set new model = " ..  modelName)
+    _G.logger:fine(nameOfModule .. ": Set new model = " ..  modelName)
     multiDeepLearning_Instances[selectedInstance].parameters.modelName = modelName
     multiDeepLearning_Instances[selectedInstance].currentModel = model
     Script.notifyEvent("MultiDeepLearning_OnNewModelFilename", multiDeepLearning_Instances[selectedInstance].parameters.modelPath .. modelName)
     Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'fullModelPath', multiDeepLearning_Instances[selectedInstance].parameters.modelPath .. modelName)
 
     local labels = MachineLearning.DeepNeuralNetwork.Model.getOutputNodeLabels(model)
-    _G.logger:info(nameOfModule .. ': Labels of model = ' .. table.concat(labels, ','))
+    _G.logger:fine(nameOfModule .. ': Labels of model = ' .. table.concat(labels, ','))
     Script.notifyEvent('MultiDeepLearning_OnNewModelLabels', table.concat(labels, ','))
     Script.notifyEvent("MultiDeepLearning_OnNewSelectedModel", modelName)
 
   else
-    _G.logger:info(nameOfModule .. ": Loading of model did not work")
+    _G.logger:warning(nameOfModule .. ": Loading of model did not work")
   end
 end
 Script.serveFunction("CSK_MultiDeepLearning.setModelByName", setModelByName)
 
 local function uploadFinished(status)
   if status == true then
-    _G.logger:info(nameOfModule .. ': New model was uploaded to the device.')
+    _G.logger:fine(nameOfModule .. ': New model was uploaded to the device.')
     local fileList = File.list(multiDeepLearning_Instances[selectedInstance].parameters.modelPath)
     if fileList ~= nil and #fileList ~= 0 then
       multiDeepLearning_Instances[selectedInstance].modelList = fileList
@@ -280,55 +280,53 @@ local function uploadFinished(status)
     end
     Script.notifyEvent("MultiDeepLearning_OnNewModelList", fileList)
   else
-    _G.logger:info(nameOfModule .. ': Error during upload of new model.')
+    _G.logger:warning(nameOfModule .. ': Error during upload of new model.')
   end
 end
 Script.serveFunction("CSK_MultiDeepLearning.uploadFinished", uploadFinished)
 
 local function downloadFromDevice(status)
-  _G.logger:info(nameOfModule .. ': Downloading model from device = ' .. tostring(status))
+  _G.logger:fine(nameOfModule .. ': Downloading model from device = ' .. tostring(status))
 end
 Script.serveFunction("CSK_MultiDeepLearning.downloadFromDevice", downloadFromDevice)
 
 local function setValidScore(value)
-  _G.logger:info(nameOfModule .. ": Set valid score to = " ..  tostring(value))
+  _G.logger:fine(nameOfModule .. ": Set valid score to = " ..  tostring(value))
   multiDeepLearning_Instances[selectedInstance].parameters.validScore = value
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'validScore', multiDeepLearning_Instances[selectedInstance].parameters.validScore)
 end
 Script.serveFunction("CSK_MultiDeepLearning.setValidScore", setValidScore)
 
 local function setViewerActive(status)
-  _G.logger:info(nameOfModule .. ": Set viewer active = " ..  tostring(status))
+  _G.logger:fine(nameOfModule .. ": Set viewer active = " ..  tostring(status))
   multiDeepLearning_Instances[selectedInstance].parameters.showImage = status
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'showImage', multiDeepLearning_Instances[selectedInstance].parameters.showImage)
 end
 Script.serveFunction("CSK_MultiDeepLearning.setViewerActive", setViewerActive)
 
 local function setRegisterEvent(event)
-  _G.logger:info(nameOfModule .. ": Set registeredEvent to = " ..  event)
+  _G.logger:fine(nameOfModule .. ": Set registeredEvent to = " ..  event)
   multiDeepLearning_Instances[selectedInstance].parameters.registeredEvent = event
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'registeredEvent', event)
 end
 Script.serveFunction("CSK_MultiDeepLearning.setRegisterEvent", setRegisterEvent)
 
 local function setForwardImage(status)
-  _G.logger:info(nameOfModule .. ": Set forwardImage to = " ..  tostring(status))
+  _G.logger:fine(nameOfModule .. ": Set forwardImage to = " ..  tostring(status))
   multiDeepLearning_Instances[selectedInstance].parameters.forwardResultWithImage = status
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'forwardResultWithImage', status)
 end
 Script.serveFunction('CSK_MultiDeepLearning.setForwardImage', setForwardImage)
 
----@param status bool
 local function setProcessWithScores(status)
-  _G.logger:info(nameOfModule .. ": Set processWithScores to = " ..  tostring(status))
+  _G.logger:fine(nameOfModule .. ": Set processWithScores to = " ..  tostring(status))
   multiDeepLearning_Instances[selectedInstance].parameters.processWithScores = status
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'processWithScores', status)
 end
 Script.serveFunction('CSK_MultiDeepLearning.setProcessWithScores', setProcessWithScores)
 
----@param status bool
 local function setSortResultByIndex(status)
-  _G.logger:info(nameOfModule .. ": Set sortResultByIndex to = " ..  tostring(status))
+  _G.logger:fine(nameOfModule .. ": Set sortResultByIndex to = " ..  tostring(status))
   multiDeepLearning_Instances[selectedInstance].parameters.sortResultByIndex = status
   Script.notifyEvent('MultiDeepLearning_OnNewImageProcessingParameter', selectedInstance, 'sortResultByIndex', status)
 end